Remove unused internal code from `servicetalk-grpc-api`
Motivation:

#1893 removed gRPC filter functionality, but left some internal code
required to support removed filters.

Modifications:

- Remove `GrpcRoutes#drainToStreamingRoutes()`;
- Remove `GrpcRouteConversions`, remove conversions from `RouteProvider`;
- Make `RouteProvider` class private;
- Remove unused `GrpcUtils#newResponse` overload;

Result:

No unused code.
